Moved slider package
