+++ app/src/main/AndroidManifest.xml
@@ -104,7 +104,7 @@
             android:label="@string/radio_button" />
 
         <activity
-            android:name=".activity.SliderActivity"
+            android:name=".components.slider.SliderActivity"
             android:label="@string/slider" />
 
         <activity
